Command split
Moved commands from index to bot_modules folder and partly rewrote them
