+++ bot_modules/commands.json
@@ -1,0 +1,43 @@
+{
+    "help": {
+        "description": "List of all bot commands",
+        "alias": ["help", "h", "commands"]
+    },
+    "join": {
+        "description": "let the bot join [your/the told] channel",
+        "alias":  ["j", "connect"]
+    },
+    "leave": {
+        "description": "bot leaves your channel",
+        "alias":  ["quit", "q", "l"]
+    },
+    "sound": {
+        "description": "Sound control commands\n Use:\nsend a sound with the command dplay or add [Title (optional)]",
+        "subcommands": {
+            "play": {
+                "description": "Plays a sound",
+                "alias": ["p"]
+            },
+            "dplay": {
+                "description": "Directly plays your sound",
+                "alias": ["d"]
+            },
+            "delete": {
+                "description": "Delete sound from the board",
+                "alias": ["r", "rm", "del"]
+            },
+            "list": {
+                "description": "List your sounds",
+                "alias": ["l"]
+            },
+            "add": {
+                "description": "Add sound to the board",
+                "alias": ["a"]
+            }
+        }
+    },
+    "stop": {
+        "description": "Stop the bot.",
+        "alias": []
+    }
+}
