+++ bot_modules/test.js
@@ -1,0 +1,3 @@
+module.exports.run = async (bot, message, args) => {
+    console.log("test.js loaded.");
+}
