+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "soundbot",
   "version": "1.0.0",
-  "description": "Discord SOundboard Bot",
+  "description": "Discord Soundboard Bot",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
